types 폴더 타입 모아두기 리펙토링, 필요없는 주석 삭제
