+++ src/components/auth/SignInForm.tsx
@@ -24,7 +24,6 @@ export default function SignInForm() {
                     <div className="flex items-center justify-between">
                         <label htmlFor="password" className="block text-sm font-medium leading-6 text-gray-900">비밀번호</label>
                         <div className="text-sm">
-                            {/* <a href="#" className="font-semibold text-indigo-600 hover:text-indigo-500">Forgot password?</a> */}
                         </div>
                     </div>
                     <div className="mt-2">
