+++ src/components/auth/SignUpForm.tsx
@@ -1,5 +1,4 @@
 import { useAuthForm } from '../../hooks';
-
 export default function SignUpForm() {
     const { email, password, isFormValid, handleEmailChange, handlePasswordChange, handleSignUpSubmit } = useAuthForm();
 
@@ -23,7 +22,6 @@ export default function SignUpForm() {
                         <div className="flex items-center justify-between">
                             <label htmlFor="password" className="block text-sm font-medium leading-6 text-gray-900">비밀번호</label>
                             <div className="text-sm">
-                                {/* <a href="#" className="font-semibold text-indigo-600 hover:text-indigo-500">Forgot password?</a> */}
                             </div>
                         </div>
                         <div className="mt-2">
