+++ src/features/index.ts
@@ -1,3 +1,2 @@
 export * from "./auth/validation";
 export {todoReducer, addTodo, deleteTodo, toggleTodo, updateTodo, initTodo} from "./todo/todo";
-export {type ActionType} from "./todo/todo";
