+++ src/features/todo/todo.ts
@@ -1,15 +1,8 @@
-import {TodoType} from "../../types";
+import {ActionType, AddTodoAction, DeleteTodoAction, InitTodoAction, TodoType, ToggleTodoAction, UpdateTodoAction} from "../../types";
 
 const TOGGLE_TODO = "TOGGLE_TODO";
 const ADD_TODO = "ADD_TODO";
 const DELETE_TODO = "DELETE_TODO";
-
-export type ActionType =
-  | {type: "TOGGLE_TODO"; id: number}
-  | {type: "ADD_TODO"; todo: TodoType}
-  | {type: "DELETE_TODO"; id: number}
-  | {type: "INIT_TODO"; todos: TodoType[]}
-  | {type: "UPDATE_TODO"; todo: TodoType};
 
 export function todoReducer(state: TodoType[], action: ActionType): TodoType[] {
   switch (action.type) {
@@ -27,28 +20,6 @@ export function todoReducer(state: TodoType[], action: ActionType): TodoType[] {
       return state;
   }
 }
-
-type AddTodoAction = {
-  type: "ADD_TODO";
-  todo: TodoType;
-};
-type ToggleTodoAction = {
-  type: "TOGGLE_TODO";
-  id: number;
-};
-type DeleteTodoAction = {
-  type: "DELETE_TODO";
-  id: number;
-};
-
-type UpdateTodoAction = {
-  type: "UPDATE_TODO";
-  todo: TodoType;
-};
-type InitTodoAction = {
-  type: "INIT_TODO";
-  todos: TodoType[];
-};
 
 export function addTodo(todo: TodoType): AddTodoAction {
   return {
