+++ src/types/index.ts
@@ -1,1 +1,11 @@
-export type { default as TodoType } from "./todo";
+export type {
+  TodoType,
+  ActionType,
+  AddTodoAction,
+  ToggleTodoAction,
+  DeleteTodoAction,
+  UpdateTodoAction,
+  InitTodoAction,
+  UseTodoApiHandlerHookType,
+  UseTodoReducerHookType,
+} from "./todo";
